Extract `write_event`
